backend/requirements.txt
requests
social-auth-app-django
django-cors-headers
django-cryptography